senario1/python/requirements.txt
flask
flask-sqlalchemy
flask-smorest
python-dotenv
marshmallow
mysql-connector-python
deepdiff
dogpile.cache

